Use size_t instead of uint

+++ include/rhc_vec.h
@@ -4,7 +4,7 @@
 #include "rhc_misc.h"
 
 typedef struct{
-  uint size;
+  size_t size;
   double *elem;
 } _vec_t;
 
@@ -21,8 +21,8 @@ typedef _vec_t * vec_t;
 vec_t vec_set_elem_list(vec_t v, ... );
 
 /* create, destroy a vector */
-vec_t vec_create(uint size);
-vec_t vec_create_list(uint size, ... );
+vec_t vec_create(size_t size);
+vec_t vec_create_list(size_t size, ... );
 void vec_destroy(vec_t v);
 
 #endif /* __RHC_VEC_H__ */
